use models as source for db
